+++ CHANGELOG.md
@@ -7,12 +7,16 @@ timeline: true
 
 ###### For now there is not schedule, make releases as soon as possible and when i have a time
 
-#### 0.4.10 `2020-03-14`
+#### 0.4.10, 0.4.11 `2020-03-14`
 
 - `NEW!` AutoComplete component added;
+- `NEW!` New icons added;
+- Icon component updated (custom icon setting added);
 - Tag component updated;
 - Select component updated;
 - Dropdown component updated;
+- Checkbox component updated,
+- Table component updated
 
 #### 0.4.8, 0.4.9 `2020-03-11`